+++ R/cells.R
@@ -24,9 +24,9 @@ setMethod("cells", signature(x="SpatRaster", y="numeric"),
 
 
 setMethod("cells", signature("SpatRaster", "SpatVector"), 
-	function(x, y, method="simple", weights=FALSE, touches=is.lines(y)) {
+	function(x, y, method="simple", weights=FALSE, exact=FALSE, touches=is.lines(y)) {
 		method = match.arg(tolower(method), c("simple", "bilinear"))
-		d <- x@ptr$vectCells(y@ptr, touches[1], method[1], weights[1] ) 
+		d <- x@ptr$vectCells(y@ptr, touches[1], method[1], weights[1], exact[1] ) 
 		if (geomtype(y) == "points") {
 			d <- matrix(d, nrow=nrow(y), byrow=TRUE)
 			d <- cbind(1:nrow(y), d)
@@ -40,7 +40,7 @@ setMethod("cells", signature("SpatRaster", "SpatVector"),
 			return (d)
 		}
 		cn <- c("ID", "cell")
-		if (weights[1]) {
+		if (weights[1] || exact[1]) {
 			d <- matrix(d, ncol=3)
 			cn <- c(cn, "weights")
 		} else {