+++ R/coerce.R
@@ -40,14 +40,14 @@ setMethod("as.raster", signature(x="SpatRaster"),
  
  
 setMethod("as.polygons", signature(x="SpatRaster"), 
-	function(x, trunc=TRUE, dissolve=TRUE, values=TRUE, extent=FALSE) {
+	function(x, trunc=TRUE, dissolve=TRUE, values=TRUE, na.rm=TRUE, extent=FALSE) {
 		p <- methods::new("SpatVector")
 		if (extent) {
 			p@ptr <- x@ptr$dense_extent()
 			x <- messages(x)
 		} else {
 			opt <- spatOptions()
-			p@ptr <- x@ptr$as_polygons(trunc[1], dissolve[1], values[1], TRUE, opt)
+			p@ptr <- x@ptr$as_polygons(trunc[1], dissolve[1], values[1], na.rm[1], opt)
 			x <- messages(x)
 			if (values) {
 				ff <- is.factor(x)