+++ src/distance.cpp
@@ -471,3 +471,4 @@ void nearest_lonlat_self(std::vector<long> &id, std::vector<double> &d, std::vec
 }
 
 
+